requirements.txt
Django==4.2.7
django-extensions==3.2.3
Pillow==10.1.0
django-crysp-forms
crispy-bootstrap5
django-bootstrap-modal-forms
django-widget-tweaks